+++ CHANGES.md
@@ -5,6 +5,8 @@
 ##### Additions :tada:
 
 - Added support for rendering glTFs with line primitives.
+- Added `UCesiumGoogleMapTilesRasterOverlay`.
+- Added `AssetOptions` property to `UCesiumIonRasterOverlay`.
 
 ### v2.19.1 - 2025-09-02